Dinesh-Kumar/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fpdf==1.7.2
